#589: Move mason tasks into main file
